fix: pin 기능들 훅으로 분리
